refactor: Composable関数として切り出しました
